Added GUI Classes
The Main class will now run as an Applet, allowing the user to select
product, release, and report type. This will bring up the GUI of the
report, which pulls dummy data from the CommitmentRatioReport and
ReleaseForecastReport classes. All remaining changes to the project
should only have to occur in those two classes.